+++ nessai/nestedsampler.py
@@ -118,34 +118,39 @@ class NestedSampler:
         Keyword arguments passed to the flow proposal class
     """
 
-    def __init__(self, model, nlive=1000, output=None,
-                 stopping=0.1,
-                 max_iteration=None,
-                 checkpointing=True,
-                 checkpoint_on_training=False,
-                 resume_file=None,
-                 seed=None,
-                 n_pool=None,
-                 plot=True,
-                 proposal_plots=True,
-                 prior_sampling=False,
-                 analytic_priors=False,
-                 maximum_uninformed=1000,
-                 uninformed_proposal=None,
-                 uninformed_acceptance_threshold=None,
-                 uninformed_proposal_kwargs=None,
-                 flow_class=None,
-                 flow_config=None,
-                 training_frequency=None,
-                 train_on_empty=True,
-                 cooldown=100,
-                 memory=False,
-                 reset_weights=False,
-                 reset_permutations=False,
-                 retrain_acceptance=True,
-                 reset_acceptance=False,
-                 acceptance_threshold=0.01,
-                 **kwargs):
+    def __init__(
+        self,
+        model,
+        nlive=2000,
+        output=None,
+        stopping=0.1,
+        max_iteration=None,
+        checkpointing=True,
+        checkpoint_on_training=False,
+        resume_file=None,
+        seed=None,
+        n_pool=None,
+        plot=True,
+        proposal_plots=False,
+        prior_sampling=False,
+        analytic_priors=False,
+        maximum_uninformed=None,
+        uninformed_proposal=None,
+        uninformed_acceptance_threshold=None,
+        uninformed_proposal_kwargs=None,
+        flow_class=None,
+        flow_config=None,
+        training_frequency=None,
+        train_on_empty=True,
+        cooldown=200,
+        memory=False,
+        reset_weights=False,
+        reset_permutations=False,
+        retrain_acceptance=True,
+        reset_acceptance=False,
+        acceptance_threshold=0.01,
+        **kwargs
+    ):
 
         logger.info('Initialising nested sampler')
 
@@ -338,7 +343,8 @@ class NestedSampler:
             priors rather than using rejection sampling.
         maximum_uninformed : {False, None, int, float}
             Maximum number of iterations before switching to FlowProposal.
-            If None, no max is set. If False uninformed sampling is not used.
+            If None, two times nlive is used. If False uninformed sampling is
+            not used.
         uninformed_acceptance_threshold :  float or None:
             Threshold to use for uninformed proposal, once reached proposal
             method will switch. If None acceptance_threshold is used if
@@ -348,13 +354,13 @@ class NestedSampler:
         """
         if maximum_uninformed is None:
             self.uninformed_sampling = True
-            self.maximum_uninformed = np.inf
+            self.maximum_uninformed = 2 * self.nlive
         elif not maximum_uninformed:
             self.uninformed_sampling = False
             self.maximum_uninformed = 0
         else:
             self.uninformed_sampling = True
-            self.maximum_uninformed = maximum_uninformed
+            self.maximum_uninformed = float(maximum_uninformed)
 
         if uninformed_acceptance_threshold is None:
             if self.acceptance_threshold < 0.1: